HiddenFiles.jl Documentation

HiddenFiles.jl Documentation

Adding HiddenFiles.jl

Documentation

Adding HiddenFiles.jl

julia> using Pkg
julia> Pkg.add("HiddenFiles")    Updating registry at `~/.julia/registries/General.toml`
   Resolving package versions...
   Installed HiddenFiles ─ v0.1.3
    Updating `~/work/HiddenFiles.jl/HiddenFiles.jl/docs/Project.toml`
  [d01c2003] ~ HiddenFiles v0.1.3 `~/work/HiddenFiles.jl/HiddenFiles.jl` ⇒ v0.1.3
    Updating `~/work/HiddenFiles.jl/HiddenFiles.jl/docs/Manifest.toml`
  [d01c2003] ~ HiddenFiles v0.1.3 `~/work/HiddenFiles.jl/HiddenFiles.jl` ⇒ v0.1.3
Precompiling project...
    314.5 ms  ✓ HiddenFiles
  1 dependency successfully precompiled in 1 seconds. 52 already precompiled.
  1 dependency precompiled but a different version is currently loaded. Restart julia to access the new version. Otherwise, loading dependents of this package may trigger further precompilation to work with the unexpected version.

Documentation

HiddenFiles.ishidden — Function

ishidden(f::AbstractString)

Check if a file or directory is hidden.

On Unix-like systems, a file or directory is hidden if it starts with a full stop/period (U+002e). On Windows systems, this function will parse file attributes to determine if the given file or directory is hidden.

Note

Directory references (i.e., . or ..) are always hidden. To check if the underlying path is hidden, you should run ishidden on its realpath.

Note

On Unix-like systems, in order to correctly determine if the file begins with a full stop, we must first expand the path to its real path.

Note

On operating systems deriving from BSD (i.e., *BSD, macOS), this function will also check the st_flags field from stat to check if the UF_HIDDEN flag has been set.

Note

On macOS, any file or directory within a package or a bundle will also be considered hidden.

Note

There may be some UNIX-specific system directories in macOS that are not yet classified as hidden (#18).

Note

Mount points on ZFS are not yet classified as hidden (#20).